x
Bench: 638619

+++ src/tt.h
@@ -5,7 +5,7 @@ struct TEntry {
 	U64 key;
 	int score;
 	uint16_t move;
-	uint16_t age;
+	uint8_t age;
 	uint8_t depth;
 	Flag flag;
 }__attribute__((packed));
